Update weak_ai.py
In this commit I added the error margin class this is created from an average type class with the average rgb values. It is this error margin class which gets compared against the tested images. In the future I plan to use the minimum and maximum components of the average_type class in order to create a range of possible matches and improve the script.