Add files via upload
These are the two photos which I tested against the ml. The Iron returned scores upwards of 75%, while the red square returned less than 2%. I believe the reason for which the actual picture of iron did not score higher is because I only used five testing images for the script to learn from.